feat: add user email to localStorage key
